Method calculating total freights value for a given forwarder was added

+++ src/main/java/com/sojerdev/forwarder/forwarder/ForwarderServiceImpl.java
@@ -1,19 +1,27 @@
 package com.sojerdev.forwarder.forwarder;
 
+import com.sojerdev.forwarder.carriage.freight.Freight;
+import com.sojerdev.forwarder.carriage.freight.FreightService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
 public class ForwarderServiceImpl implements ForwarderService {
 
     private final ForwarderRepository forwarderRepository;
+    private final FreightService freightService;
 
     @Autowired
-    public ForwarderServiceImpl(ForwarderRepository forwarderRepository) {
+    public ForwarderServiceImpl(ForwarderRepository forwarderRepository, FreightService freightService) {
         this.forwarderRepository = forwarderRepository;
+        this.freightService = freightService;
     }
 
     @Override
@@ -43,6 +51,31 @@ public class ForwarderServiceImpl implements ForwarderService {
 
     @Override
     public void deleteById(int id) {
+        if (findById(id) == null) {
+            throw new RuntimeException("Forwarder with id - " + id + " was not found.");
+        }
         forwarderRepository.deleteById(id);
+    }
+
+    @Override
+    public ResponseEntity<Object> getTotalValue(int id) {
+        if (findById(id) == null) {
+            throw new RuntimeException("Forwarder with id - " + id + " was not found.");
+        }
+        List<Freight> freights = freightService.findAll();
+        int totalValue = 0;
+        int numOfFreights = 0;
+
+        for (Freight freight: freights) {
+            if (freight.getCarriage().getForwarder().getId() == id) {
+                totalValue += freight.getValue();
+                numOfFreights++;
+            }
+        }
+        Map<String, Integer> response = new HashMap<>();
+        response.put("freights", numOfFreights);
+        response.put("value", totalValue);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
